combine feedUrl and feed package
