+++ src/github.com/getlantern/flashlight/feed/feed_test.go
@@ -1,4 +1,4 @@
-package lantern
+package feed
 
 import (
 	"testing"
